+++ src/main.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 )
 
-const version = "0.2.0"
+const version = "0.2.1"
 
 func main() {
 	args := os.Args[1:]